Infer enum variant payload
